added instant video display after creation

+++ CommentsViewController.swift
@@ -271,42 +271,69 @@ class CommentsViewController: UIViewController, UITableViewDelegate, UITableView
                         cell.timestampLabel.text = timeSinceTimeStamp(comment!.createdAt!)
                     }
                     
-                    let videoFile = comment!["video"] as! PFFile
-                    videoFile.getDataInBackgroundWithBlock {
-                        (videoData, error) -> Void in
-                        if error == nil {
-                            var path = "\(self.documentPath)/\(indexPath.row).mp4"
-                            if (NSFileManager.defaultManager().fileExistsAtPath(path)) {
-                                NSFileManager.defaultManager().removeItemAtPath(path, error: nil)
-                            }
+
+                    if indexPath.row == 0 && videoJustCreated == true {
+                        var movieURL = NSURL(fileURLWithPath: self.videoPath!)
+                        cell.player = AVPlayer(URL: movieURL)
+                        
+                        
+                        cell.playerLayer = AVPlayerLayer(player: cell.player!)
+                        cell.playerLayer!.frame = CGRectMake(0, 0, self.screenSize.width, self.screenSize.width)
+                        cell.playerLayer!.videoGravity = AVLayerVideoGravityResizeAspect
+                        cell.playerLayer!.needsDisplayOnBoundsChange = true
+                        
+                        cell.contentView.layer.insertSublayer(cell.playerLayer!, atIndex: 0)
+                        cell.contentView.layer.needsDisplayOnBoundsChange = true
+                        
+                        if self.cellCompletelyOnScreen(indexPath){
+                            self.playingVideoCell = cell
+                            self.playingVideoCell?.playButtonIconImageView.hidden = true
+                            self.playingVideoCell!.player!.play()
+                            self.playingVideoCell!.player!.actionAtItemEnd = .None
                             
-                            videoData!.writeToFile(path, atomically: true)
-                            println("File now at \(path)")
-                            var movieURL = NSURL(fileURLWithPath: path)
-                            cell.player = AVPlayer(URL: movieURL)
                             
-                            
-                            cell.playerLayer = AVPlayerLayer(player: cell.player!)
-                            cell.playerLayer!.frame = CGRectMake(0, 0, self.screenSize.width, self.screenSize.width)
-                            cell.playerLayer!.videoGravity = AVLayerVideoGravityResizeAspect
-                            cell.playerLayer!.needsDisplayOnBoundsChange = true
-                            
-                            cell.contentView.layer.insertSublayer(cell.playerLayer!, atIndex: 0)
-                            cell.contentView.layer.needsDisplayOnBoundsChange = true
-                            
-                            if self.cellCompletelyOnScreen(indexPath){
-                                self.playingVideoCell = cell
-                                self.playingVideoCell?.playButtonIconImageView.hidden = true
-                                self.playingVideoCell!.player!.play()
-                                self.playingVideoCell!.player!.actionAtItemEnd = .None
+                            NSNotificationCenter.defaultCenter().addObserver(self, selector: "restartVideoFromBeginning", name: AVPlayerItemDidPlayToEndTimeNotification, object: self.playingVideoCell!.player!.currentItem)
+                        }
+                    } else {
+                        let videoFile = comment!["video"] as! PFFile
+                        videoFile.getDataInBackgroundWithBlock {
+                            (videoData, error) -> Void in
+                            if error == nil {
+                                var path = "\(self.documentPath)/\(indexPath.row).mp4"
+                                if (NSFileManager.defaultManager().fileExistsAtPath(path)) {
+                                    NSFileManager.defaultManager().removeItemAtPath(path, error: nil)
+                                }
+                                
+                                videoData!.writeToFile(path, atomically: true)
+                                println("File now at \(path)")
+                                var movieURL = NSURL(fileURLWithPath: path)
+                                cell.player = AVPlayer(URL: movieURL)
                                 
                                 
-                                NSNotificationCenter.defaultCenter().addObserver(self, selector: "restartVideoFromBeginning", name: AVPlayerItemDidPlayToEndTimeNotification, object: self.playingVideoCell!.player!.currentItem)
+                                cell.playerLayer = AVPlayerLayer(player: cell.player!)
+                                cell.playerLayer!.frame = CGRectMake(0, 0, self.screenSize.width, self.screenSize.width)
+                                cell.playerLayer!.videoGravity = AVLayerVideoGravityResizeAspect
+                                cell.playerLayer!.needsDisplayOnBoundsChange = true
+                                
+                                cell.contentView.layer.insertSublayer(cell.playerLayer!, atIndex: 0)
+                                cell.contentView.layer.needsDisplayOnBoundsChange = true
+                                
+                                if self.cellCompletelyOnScreen(indexPath){
+                                    self.playingVideoCell = cell
+                                    self.playingVideoCell?.playButtonIconImageView.hidden = true
+                                    self.playingVideoCell!.player!.play()
+                                    self.playingVideoCell!.player!.actionAtItemEnd = .None
+                                    
+                                    
+                                    NSNotificationCenter.defaultCenter().addObserver(self, selector: "restartVideoFromBeginning", name: AVPlayerItemDidPlayToEndTimeNotification, object: self.playingVideoCell!.player!.currentItem)
+                                }
+                                
+                                
                             }
-                            
-                            
                         }
                     }
+                    
+                    
                     
                     var commentUser : PFUser = comment!["user"] as! PFUser
                     commentUser.fetchIfNeededInBackgroundWithBlock {
@@ -1206,6 +1233,10 @@ class CommentsViewController: UIViewController, UITableViewDelegate, UITableView
     }
     
     func deleteCell(cell: UITableViewCell) {
+        if playingVideoCell != nil && playingVideoCell?.player?.rate == 1.0 {
+            playingVideoCell?.player?.pause()
+            playingVideoCell?.playButtonIconImageView.hidden = false
+        }
         var deleteCellIndexPath : NSIndexPath = self.commentsTableView.indexPathForCell(cell)!
         var temporaryCommentsArray = NSMutableArray(array: comments)
         temporaryCommentsArray.removeObjectAtIndex(deleteCellIndexPath.row)
